updates to methods for getting system properties
git-svn-id: https://findbugs.googlecode.com/svn/trunk@10884 eae3c2d3-9b19-0410-a86e-396b6ccb6ab3

+++ findbugs/src/java/edu/umd/cs/findbugs/detect/FindInconsistentSync2.java
@@ -112,7 +112,7 @@ public class FindInconsistentSync2 implements Detector {
 	 * intentional synchronization.
 	 */
 	private static final int MIN_SYNC_PERCENT =
-			SystemProperties.getInteger("findbugs.fis.minSyncPercent", 50).intValue();
+			SystemProperties.getInt("findbugs.fis.minSyncPercent", 50);
 
 	/**
 	 * Bias that writes are given with respect to reads.
